+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+* Appbar can be animated by providing a ScrollController to it.
+#### Breaking changes:
+Appbar:
+- `background` is replaced with `pathToBackground`.
 # 0.1.1
 * More pub points
 # 0.1.0